css/css-overscroll-behavior/parsing/overscroll-behavior-y-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overscroll Behavior: getComputedValue().overscrollBehaviorY</title>
<link rel="help" href="https://drafts.csswg.org/css-overscroll-behavior-y/#propdef-overscroll-behavior-y">
<meta name="assert" content="overscroll-behavior-y computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("overscroll-behavior-y", "contain");
test_computed_value("overscroll-behavior-y", "none");
test_computed_value("overscroll-behavior-y", "auto");
</script>
</body>
</html>
